separate tests into domain and command tests
